BinaryDecompiler: Include toGLSL functions from HLSLCrossCompiler
This will form the basis of an experiment to see if we can make
HLSLCrossCompiler generate HLSL rather than just using it to decode the
shaders. This brings in the functions in HLSLCrossCompiler involved in
generating GLSL structured buffer loads and everything they depend on.

The dependencies probably brings in a lot more than we need, but if this
is successful we may either strip it down to what is necessary, or
consider bringing in more of HLSLCrossCompiler to bring our version
closer to parity with upstream to lower our maintenance burden.