BinaryDecompiler: Fix incorrect use of malloc & remove previous hack
Comparing the Binary Decompiler project with the upstream
HLSLCrossCompiler I spotted some code we had changed to fix a crash, but
our fix involved a very suspect hack where some C++ objects were
constructed and copied over the top of another.

A closer inspection revealed the actual culprit was that the
ShaderVarType structure contains C++ strings, so it needs to be
allocated using new, not malloc so that the contained C++ objects are
properly constructed. Most mallocs from the original HLSLCrossCompiler
project had already been replaced with new when Chiri ported it to C++,
but this one was clearly missed.

Reverting the previous hack makes the HLSL decompiler regression test
suite crash randomly, but once this fix is applied the crashes go away
again, giving me a high degree of confidence that this is the correct
fix.