BinaryDecompiler: Replace compiler warning suppression with cast
The compiler warns here because asVars has been converted into a C++
data structure and we use size() to get the length, compared to the
original C code which stored the length separately. Since this line
already differs from the original HLSLCrossCompiler code, we may as well
do the cast inline and remove the pragma and comment to keep our diff to
the original code as small as possible, and avoid the risk of masking
other bugs.