cmd_Decompiler: Add --copy-reflection option
This option passes an already compiled shader binary to Flugan's
assembler to use the reflection, signature, stat and any other esoteric
sections from instead of assembling the signature sections ourselves.

This allows us to use the original behaviour of Flugan's assembler (or
the current behaviour with assemble_signature_comments disabled) from
the command line with the caveat that we must have an appropriate shader
binary with the sections we want to use. It is also currently the only
way to assemble shaders from the command line with reflection
information, as ordinarily we do not assemble the comments corresponding
to that section.

My immediate goal is to use this to assemble the shaders under
TestShaders/GameExamples to create test cases for the HLSL decompiler,
but restoring the reflection information that is currently lost when
assembling these.