Add script to reconstruct binary shaders from dumped HLSL
In order to perform regression tests on the HLSL decompiler we need to
start with a pristine shader - either one that we have written and
compiled ourselves, or a precompiled shader from a game.

We already have a collection of shaders from games intended to be used
as test cases, but these are shaders that have already been through the
decompiler, and therefore are not quite the same as running the
decompiler over the original shader. Ideally we would have saved these
using export_binary=1, but in the absence of that this script attempts
to recreate something as close to the original shader binary as
possible.

If the shader was dumped with export_hlsl=2/3 we pull out the original
assembly block and assemble that. However, the assembler does not
assemble reflection information and we really need that preserved to
adequately test the decompiler. Therefore, we also compile the HLSL in
these shaders and use the reflection information from those combined
with the bytecode from the assembler to reconstruct the original binary
as close as possible with such an automated process.

This may not be perfect - if a test shader fails to compile it will fail
outright, or if the instructions that access a resource are missing the
HLSL compiler will optimise out the referenced resources. We could of
course get closer by extending the assembler to process the resource
binding comments, but that's a considerable amount of work for something
we don't otherwise need.