Turn example shaders into decompiler test cases
This turns every HLSL shader in the GameExamples directory where we were
able to reconstruct a binary shader (by assembling the assembly version
and combining it with reflection information from compiling the HLSL
example) into a potential test case for the HLSL decompiler.

For now the .chk files are just copies of the existing HLSL text* - these
will need to be regenerated with a recent version of the decompiler to
account for any fixes/changes to the decompiler since these were dumped
and to remove any manual edits, but I want that in a separate commit so
that the diff will clearly show these changes.

* with NULL bytes removed to ensure that git will treat them as text.

The reconstructed binary shaders are not included, on the theory that
they are trivial to reconstruct from the example HLSL text on demand and
would pollute our git repository with unnecessary binary blobs.