Update HLSL Decompiler test suite to use GameExample shaders
This adds a run_game_example_tests.sh script which will run over all the
.hlsl.chk files found under the GameExamples directory, reconstruct
their binary shaders (if necessary) then run the HLSL decompiler over
them, compare the result to the .chk file and attempt to recompile the
decompiled shader.

At the moment, none of the shaders match their .chk files so all tests
fail, but we will update those in a following patch. Notably a number of
these example shaders also fail the recompilation stage that we should
look at later.