HLSL Decompiler Test Suite: Handle bad shader binaries
The GameExamples/DR3/cc5538d28f8fd45e-vs.shdr reconstructed shader is
causing FXC to hang when disassembling it. This is something we will
need to investigate in case it signifies a bug in our assembler (not so
worried if it is just a reconstruction issue), but we don't want it to
block the rest of the tests.

This adds a 5 second timeout when calling fxc to get the shader model.