Update GameExamples .chk files from cmd_Decompiler 1.3.2
This regenerates the game example shader test case test files based on
the decompiler shipped with cmd_Decompiler 1.3.2 (being the most recent
version shipped on github) and fxc shipped with the Windows 8.0 SDK
(being the SDK we officially support) - the intent is to make this
commit as reproducible as possible in case it reveals issues that need
to be investigated later.

This commit was essentially:

    export FXC=.../Windows Kits/8.0/bin/x64/fxc.exe

    # Next version of cmd_Decompiler required for shader binary
    # reconstruction. --update-chk is _not_ used in this step:
    export CMD_DECOMPILER=.../cmd_Decompiler-1.3.12.exe
    ./run_game_example_tests.sh

    # Latest officially released version of cmd_Decompiler used to
    # update the chk files. The reconstructed files will be used from
    # above:
    export CMD_DECOMPILER=.../cmd_Decompiler-1.3.2.exe
    ./run_game_example_tests.sh --update-chk

I have not verified every change this makes as my immediate goal is to
give me a baseline as the decompiler stands at this moment.

The diff in this commit may also be useful to spot regressions in the
decompiler, however since the decompiler has evolved considerably since
most of these example shaders were added the diff will be quite large
and should mostly be quite benign.