Update test case .chk files for mixed type buffer fix
Now that the assembler properly handles these shaders we have valid
reconstructed shaders to test the HLSL decompiler, so like the others
update the .chk files to give us a baseline for regression testing from
this point on.

Notably the decompiler does not appear to handle these mixed type
buffers properly, so these tests still fail the recompilation test, but
that's a separate issue and we still need this baseline regardless.