BinaryDecompiler: Fix memory leak on Members & kill emplace_back() as broken
The "Members" structure is allocated and never released, creating a
memory leak. As far as I can tell this is still the case in the original
HLSLCrossCompiler project as well, but they may be relying on the OS to
clean it up if they are not intended to be used in a long running
process.

Since our BinaryDecompiler has been ported to C++ add a destructor to
properly release the memory.

The emplace_back() used to fix the previous problem of the structs being
copied and invalidating pointers to them did seem fix that problem in
our test cases, but is actually still problematic as during the vector
resize process it can copy or move the contained structures, once again
potentially exposing us to invalid pointer crashes and calling this
newly added destructor prematurely releasing the Members array.

Instead, use resize() on the vector to allocate all the variables ahead
of time so that they will not move and the destructors won't be
prematurely called. This is in fact closer to how the original
HLSLCrossCompiler works as it also allocates these in advance.