Copy on mark: Add bad HLSL as comment when falling back to assembly
When working with assembly shaders, it can still be nice to refer to the
HLSL versions to check things like variable names. Since we can now dump
assembly via copy on mark we could be in a situation where we might not
have ready access to the HLSL version of the shader.

To help with this use case, this patch will add the contents of the HLSL
shader in a comment at the bottom of assembly shaders when both HLSL and
assembly copy on mark are enabled, and the HLSL shader failed to
compile.