run_game_example_tests: Add mode to run specific test cases
This allows the specific test cases to be passed to the script to run
rather than running every test. This is useful when iterating on one
specific test, or to create a .chk file for a new test case that
wouldn't be picked up in the default mode.

It will handle reconstructing the binary shaders if necessary, including
cases where both the stripped or unstripped test cases have been
specified. If run without any arguments it will run all test cases as
before.