HLSL Decompiler: Fix structured buffer definition with embedded structs
Structured buffer definitions can contain other embedded structures,
which looks like:

// Resource bind info for instanceData
// {
//
//   struct SImpostorInstance
//   {
//
//       struct SPackedMatrix
//       {
//
//           uint rot[5];               // Offset:    0
//           float3 pos;                // Offset:   20
//
//       } m_treeToWS;                  // Offset:    0
//       uint m_tint;                   // Offset:   32
...

We also see variants on this where the embedded type name has been
replaced with a generic identifier:

...
//       struct SImpostorInstance::<unnamed>
...

The TestShaders/structured_buffers.hlsl shows that multiple unnamed
structures at the same level will have identical placeholder names.

The difference between these two cases is whether the embedded structure
was declared inline or not in the original HLSL code - if it was
declared inline it cannot have a type name, so it gets the ::<unnamed>
variant, while if it was declared separately it will have a valid type
name we see here.

Since the structure we are pulling out is closer to the inline version
regardless of the original we need to strip the embedded structure type
name to generate valid HLSL. This commit does that, and also cleans up
the whitespace surrounding these.