HLSL Decompiler: Handle multiple StructuredBuffers of the same type
If a shader uses multiple structured buffers with the same struct type
they will be duplicated in the assembly, but when we convert them back
to HLSL we need to ensure that each type is defined only once - keep
track of which struct type names have been used and skip defining any
duplicates.

We don't verify that the contents of the types actually match, as we
can probably assume that the HLSL compiler would have forbidden that
possibility, and if there is some situation where it does happen we can
deal with it then.