HLSL Decompiler: Fix raw buffers being mistaken for StructuredBuffers
GameExamples/JC3/f064c0539aaa5254-vs contains examples of several
different types of buffers - typed, structued and raw:

// Name                                 Type  Format         Dim Slot Elements
// ------------------------------ ---------- ------- ----------- ---- --------
// PositionInfo                      texture  struct         r/o    0        1
// IndexBuffer                       texture    uint         buf    3        1
// VertexBuffer                      texture    byte         r/o    4        1

The decompiler was incorrectly treating "VertexBuffer" (not being
accessed as a vertex buffer) as a StructuredBuffer by matching "r/o" in
the Dim column.  Change the decompiler to identify StructuredBuffers and
raw buffers by the format column.

The r/w versions of these have type=uav, which is not yet supported.