HLSL Decompiler: Fix hex constant literals being damaged by applySwizzle
A curiosity showed up on the vs2017 branch where a number of values that
had been decompiled as 0 on the master branch were decompiled as 65535
or 2139095040 (0x7f800000, or inf). Examination of the shaders in
question showed that these were hex literals that were being damaged on
the master branch, and a curious change to how scanf("%f") handles "0x"
resulted in the vs2017 branch getting closer to what these should be by
dumb luck.

In fact, this particular bug is noted in the comments of
GameExamples/FC4/9cc9ea03feed4639-vs_replace.txt and other shaders.

This commit updates applySwizzleLiteral to handle hex literals
explicitly and pass them on as intN(0x..., 0x...), etc. for instructions
that operate on integers. The 'and', 'ine' and 'ieq' instructions have
been updated to indicate they operate on integers since they are all
bitwise instructions.