HLSLCrossCompiler: Fix NULL pointer dereference on dcl_uav_structured
The following commit introduced a NULL pointer dereference that can
crash the HLSL cross compiler if it processes a shader that includes an
RWStructuredBuffer that has had its reflection information stripped:

    commit 0659b1f3dfe3c2617c0b6d5b96d54017dcf9cae9
    Author: James Jones <james.jones.jmj@gmail.com>
    Date:   Fri Jan 17 12:21:23 2014 +0000

        Fixes image_load_store and shader_storage_buffer_object features

This restores the if block to make the processing conditional on the
reflection information being found.

A version of this patch suitable for pushing upstream is here:
https://github.com/DarkStarSword/HLSLCrossCompiler/tree/rw_struct_buf_crash_fix

However, upstream may really need more work to not crash on stripped
shaders at all, as I don't think that was ever a valid use case for
them.