BinaryDecompiler: Apply 3DMigoto/C++ ports
Since 3DMigoto's version of the HLSLCrossCompiler has been ported to
C++, we need to make some changes to the newly imported code to compile
and run.

It might be worthwhile switching back to a C implementation of
HLSLCrossCompiler to reduce our maintenance burden, but that will be a
project for another day.