HLSL Decompiler: Add support for ld_structured (HLSLCrossCompiler Experiment)
This is an experiment in using HLSLCrossCompiler to produce the HLSL
code for a ld_structured instruction so that we don't have to deal with
all the edge cases.

At the moment this actually inserts GLSL code, but as a proof of concept
it works. The biggest issue is that the structure names have been
replaced with generic names that we will need to solve, and we also have
some potential compatibility issues as HLSLCrossCompiler appears to be
using reinterpret casts, but we can probably disable those in the short
term, and long term we might consider switching 3DMigoto to do the same,
as it solves some of the issues with casting everything to floats
(however, previous experiments have hit issues with asfloat/asint not
working as expected).