HLSL Decompiler: Manufacture structure definitions for stripped shaders
If a shader is missing reflection information we won't be able to fill
in an accurate type definition for each StructuredBuffer resource, so we
need to manufacture fake type information to take it's place. This is a
deeper concern than when a shader is missing constant buffer
information, because in practice we tend to see StructuredBuffers being
used for a wider range of data types than constant buffers, and so are
more likely to fall foul of casting issues when we are faking the types,
but the shaderhacker can always switch over to assembly in that case.

.chk files are updated. The only curiosity of note here was that one
non-stripped shader generated fake data types, but this appears to be
down to an artefact with the reconstruction process, where the HLSL
optimised out these resources and therefore didn't add reflection
information for them, but the original assembly still contained
references to them. We could probably make the decompiler deal with
this, but it's an edge case that we shouldn't need to care about in
practice, so probably better off fixing the test case:

    GameExamples/ACUnity/7080753c496de9b1-ps.hlsl.chk