Structured Buffer test case: Add no swizzle accesses
There is a distinction in HLSLCrossCompiler as to whether a source
operand has no swizzle (usually implying .xyzw) or an arbitrary four
component swizzle (also, one component swizzles are a separate case
in the DX bytecode), so we need to make sure that our test case has some
aligned 16 byte loads that will use the default .xyzw swizzle as well as
rearranged swizzles to test both cases.

We already have a three component case that covers the rearranged case,
this adds some extra aligned float4 loads to cover the no swizzle case.

This also updates the .chk files, and adds the stripped .chk files to
observe how they change in the following commits.