HLSL Decompiler: Implement ld_structured with no reflection information
This produces code for ld_structured when reflection information is
missing that will perform loads using the fake type information that we
have manufactured.

Since StructuredBuffers lack the hard alignment constraints of constant
buffers (although ignoring the 128 bit GPU cacheline alignment will have
performance impacts on the GPU), so we access these as an array of
floats rather than float4s, and we break up each load into multiple
loads for each component in the destination mask. We could potentially
recombine them using the floatN(...) construct for the shaderhacker's
reference, but the compiler is smart enough to put them back together
when it recompiles the shader.

We convert the source swizzle into a 32bit offset modifier into the
structure, handling cases for both explicit and default swizzles.