HLSL Decompiler: Change ld_structured code path decision
Now that we have the necessary updates to BinaryDecompiler we can use
the bind information for the specific register slot to determine whether
to use real or fake type information. This fixes the Witcher 3 shaders
thread group shared memory loads from trying to use non-existant
reflection information (though tgsm will need additional support), and
improves the behaviour of the ACUnity test case that only has partial
reflection info due to an artefact of the reconstruction process (i.e.
not something we expect to see in practice).