HLSL Decompiler: Fix looking for wrong reflection info for 'g' registers
This fixes the case where a 'g' register will try to use reflection
information from a 't' register of the same index. 'g' registers are for
compute shaders shared memory, and never have reflection information, so
we always want them to use the no reflection code path.

This also implements dcl_tgsm_structured and adds an example to the
compute.hlsl test case.