HLSLCrossCompiler: Fix misattributed variables following arrays in StructuredBuffers
If we have a StructuredBuffer similar to the following:

    struct foo {
    	int bar[8];
    	float baz;
    };
    StructuredBuffer<struct foo> test_struct;

Then HLSLCrossCompiler will misattribute test_struct[n].baz as being
test_struct[n].bar[8]. This is because IsOffsetInType treats all
elements of arrays except the final element as being the size of a vec4,
however this is only the case for constant buffers where the packing
rules are designed to avoid extra offset calculations when accessing
arrays and avoid reads that cross 128 bit GPU cacheline boundaries by
default.

StructuredBuffers do not share this same constraints, do not receive any
automatic padding and allow reads to cross cacheline boundaries. For
these IsOffsetInType should only use the basic size of each element in
the array without padding.