HLSLCrossCompiler: Fix variable size calculation for nested structs
When calculating the size of a nested struct we need to sum up the sizes
of all variables it contains, possibly including yet deeper nested
structs.

Split out the size calculation into a separate function that can
recursively call itself to calculate the full size of the shader
variable, including any nested structs.