HLSL Decompiler: Output variables nested within multiple struct levels
This uses the FullName field of the variable returned by
BinaryDecompiler to print out the full hierarchy leading to the located
variable. It also switches to building up the variable name string from
the FullName with the array indexing and swizzling appended.

This currently does not handle nested arrays of structs correctly.