HLSL Decompiler: Fix variable hierarchy containing arrays at any level
The FullName that BinaryDecompiler returns is insufficient as it does
not include array indexing up the heirachy. Move the processing out into
a separate function that can recursively call itself to build up the
name with any necessary array indices.

This crashes due to a bad parent pointer on top level members from
BinaryDecompiler.