BinaryDecompiler: Fix invalid Parent pointer in first level struct members
When HLSLCrossCompiler was ported to C++ as BinaryDecompiler for use in
3DMigoto, some of the C constructs were changed to similar C++
constructs. In particular, some fixed size arrays were changed to
vectors.

That in and of itself is not a problem, however due to the way vectors
of structs work, inserting a struct into the vector will actually insert
a copy. This is problematic, because BinaryDecompiler had set pointers
in member variables to the parent struct before they were inserted into
the vector, and as such once the struct was inserted into the vector the
pointers would no longer be valid, leading to crashes if we ever try to
dereference them.

This changes the code to emplace the new struct in the vector first and
initialise it using a pointer - this way the parent pointers should
remain valid. This fixes the crashes uncovered in the previous commit
and our nested structured buffer test cases are now passing.

As a bonus, this reduces the diff between BinaryDecompiler and the
current upstream HLSLCrossCompiler.