HLSLCrossCompiler: Add support for locating variables within nested struct arrays
If a struct has been nested with multiple array elements, such as:

    struct bar {
        float force_misalign;
    	struct {
    		float foo;
    		float bar;
    	} nested_struct[2];
    };
    StructuredBuffer<struct bar> struct_buf;

Then we need to account for offsets that fall within nested_struct[1] or
beyond. The contained variables foo and bar will have offsets recorded
for the first element only, so when processing nested structs we need to
determine if the offset lies somewhere in the struct, then modulus it to
contain it within the first element.