HLSL Decompiler: Implement store_structured
This refactors the heart of the ld_structured parsing into a separate
function that just translates the structure name based on the
swizzle/mask and separates it out from the code that writes it to the
destination so that it can be used for both loads and stores.

It writes the results into four strings depending on which components of
the mask (r/w) and swizzle (r/o) are in use, then the instruction
specific code can apply them as necessary.

This fixes the Witcher 3 shaders using uninitialised memory instead of g
registers and randomly failing regression tests.