HLSL Decompiler: Add support for RWStructuredBuffers and other UAV declarations
UAVs declarations are nearly identical to texture declarations, except
with a RW prefix and a 'u' register prefix instead of 't'.

This makes the existing texture declaration parsers process any type of
UAVs for the reflection case, and for the non-reflection case
dcl_uav_structured will generate fake type information similar to
dcl_resource_structured. The non-reflection case of typed UAVs is not
covered as yet.

With this change our RWStructuredBuffer test cases are now passing.