HLSLCrossCompiler: Remove tbuffer assertion
This assert appears to have been placed for the developer to check which
resource group "tbuffers" are part of. These tbuffers appear in several
of our game example test cases:

    GameExamples/DR3/cc5538d28f8fd45e-vs
    GameExamples/JC3/58212e194c551c65-ps
    GameExamples/JC3/bf629b118eb736e0-ps

And appear to use 't' registers like textures:

    // tbuffer WorldMatrix
    // {
    //   row_major float3x4 gWorldMatrix[4];// Offset:    0 Size:   192
    // }

    // Resource Bindings:
    // Name                                 Type  Format         Dim Slot Elements
    // ------------------------------ ---------- ------- ----------- ---- --------
    // WorldMatrix                       tbuffer      NA          NA  114        1

    dcl_resource_buffer (mixed,mixed,mixed,mixed) t114

    ld_indexable(buffer)(mixed,mixed,mixed,mixed) r0.xyzw, l(0, 0, 0, 0), t114.xyzw

So, translating them to RGROUP_TEXTURE appears to be correct and so we
can remove the harmless assert.