Test Framework: Output cmd_Decompiler & FXC messages to log files
If a test case is failing we may want to see the output from
cmd_Decompiler and/or fxc, but the test framework hides that by default
to keep the output clean. While it is always possible to run the
commands manually to find out what failed, this makes it more convenient
by directing the output to a log file.