Remove IniParams size cap
This removes the cap on the size of IniParams and instead sizes it
automatically based on what is actually used in the d3dx.ini, including
resizing it if the d3dx.ini is updated. If more than 256 IniParams are
used a notice will be issued in hunting mode since there may be
performance concerns, but the notice is not fatal and 3DMigoto will do
as they ask.

This rewords the key/preset override code to avoid the need to iterate
over the entire IniParams space, and instead uses maps of the index +
component pair to find the value or other structure it needs.