Command List: Add named variables independent from IniParams
This adds a new class of variables that aims to address several issues
with IniParams:

 - These variables can be given meaningful names
 - These variables are namespaced to avoid naming conflicts between
   multiple config files
 - These variables can be used in the command list without incurring the
   overhead required to update the IniParams resource on the GPU.

This effectively frees up IniParams to be used purely for cases where
the GPU does need to know a variable's value.

This follows the command list philosophy of paying overhead associated
with string parsing and map lookups during the config parsing stage
only, and during runtime the variables are accessed using only
lightweight pointers.

Variables start with a $ to distinguish them from other keywords and
must be declared in a [Variables] section and can be given an initial
value if desired that point (this section may not be final):

[Variables]
$hud_depth = 0.1

To access a variable in another namespace, the full namespace must be
used, such as:

[Constants]
$\ShaderFixes\3dvision2sbs.ini\Mode = 7

Currently all variables are global / namespaced - as yet there are no
variables local to a given command list, other types, arrays, etc.

This is not yet hooked up to the preset/override code (beyond the
ability for that code to run command lists of course).