Overrides: Fix updating IniParams after transitions completed
Once an override has been activated that updates IniParams or variables
the transition states for these will be permanently in the data
structures, causing IniParams to be unnecessarily updated every frame.

This removes the transitions from the data structures once they have
completed so that they will not be considered in subsequent frames.