Named Variables: Replace [Variables] section with global declarations in [Constants]
The introduction of the [Variables] section conceptually serves a near
identical purpose as the [Constants] section, but there are distinctions
as to what each section can do that may be unnecessarily confusing -
[Variables] can declare and initialise variables with simple floats, but
cannot initialise IniParams and cannot initialise variables with
anything other than a flot, and [Constants] can initialise variables,
but only if they were also declared in [Variables]...

This patch abolishes the [Variables] section altogether and instead aims
to grant the [Constants] command list the ability to declare variables
in its place. Variable assignments prefixed with the "global" keyword
are allocated in a first pass before the main command list is parsed.

The reason for keeping these in two separate passes is to allow an
included config file (such as 3dvision2sbs.ini) to declare a variable
(possibly initialised with a default value) and still allow the main
d3dx.ini's [Constants] list to override it for the user's preference,
throwing a warning if the appropriate config was not included.

This patch re-orders the command list parsing a little to parse
[Constants] before any other command list, preset or key override that
may need to reference the global variables. As part of this
reorganisation, failed CustomShaders are no longer erased from the
global map in case there might be any pointers to a failed shader in the
[Constants] list. Instead these skip their command list parsing which
results in them being optimised out. As a bonus this removes the error
on the run commands referring to these which had misled some people, so
that only the error on the custom shader itself will be reported.