Fix IniParams GPU resource not being updated on config reload
When we do an F10 config reload we expect the [Constants] command list
to update the IniParams resource on the GPU, but if that command list
doesn't initialise any IniParams to a non-zero value than it will skip
the update.

This performs an explicit update immediately after clearing the CPU
resource to ensure the two remain synchronised. This is a bit wasteful
since in most cases we will update the resource twice in a row, and the
alternative is setting a flag to force update_params in the [Constants]
command list, but since config reload is a cold path a little extra
overhead won't matter and I don't want to risk a regression if further
command list optimisations cause [Constants] to bail out early and not
consider update_params at all.