Command List: Remove recursion limits on if/elif/else and ShaderRegex links
If/elif/else are internally implemented as nested command lists, which
can potentially hit the command list recursion limit relatively easily
if there are too many else if clauses. The command list limit is
intended to prevent accidental circular call chains from crashing the
game, and is not intended to place a hard limit on the number of else if
clauses in an if block. Therefore, this patch stops if commands from
incrementing the recursion counter and contributing towards the limit.

It also removes ShaderRegex link commands from contributing towards the
recursion limit, because again they are not considered logically
recursive regardless of how they are implemented internally.