Command List / Frame Analysis: Improve logging of if/elif/else/endif
This improves the frame analysis log of the if and related clauses. This
adds an extra indent value for these that is separate from the recursion
level and is only used on the if / else if (true) and else cases - not
in the if / else if (false) cases. This also fixes the logging of the
else clause and adds curly brackets to take the place of the old pre /
post {} curlies from recursion.

Example of what the log may now look like:

    000045 3DMigoto pre {
    000045 3DMigoto  [Present] if $f == 1: false
    000045 3DMigoto  [Present] else if $f == 2: false
    000045 3DMigoto  [Present] else if $f == 3: true {
    000045 3DMigoto   [Present] clear = bb 1 1 0 0
    000045 3DMigoto     clearing RTV
    000045 3DMigoto  } endif
    000045 3DMigoto }