Fix debug assertion in delete while iterating pattern
Turns out trying to be smart is sometimes dumb. This form of the delete
while iterating pattern is nice for clarity because the current and next
iterators remain consistent throughout the loop, but incrementing the
next iterator in the for's next construct will point it beyond the
container's end() at the end of the final iteration, which is of no
consequence in release builds since it won't be used but it does fire a
debug assertion in debug builds.

Move the next++ inside the loop to prevent that final increment.