Command List: Implement local variables
This adds support for declaring local variables in the command lists,
like:

    local $foo = 0

The local variables are stored in the main pre command lists, and are
accessed by pointer from variable assignment commands and operands in
the pre, post and if block command lists.

These local variables are actually more like static variables in that
they are associated with a given command list and will retain their
contents between executions - this is done primarily to allow local
variables to be shared between the pre and post runs of a command list,
with the downsides being limitations in use in recursive algorithms and
potential threading issues. We can look at solutions for these if they
are needed in the future (e.g. using a stack for recursive algorithms
and locating variables in a per-thread/context buffer for the later),
but this should cover the more common requirements for local variables.

A scope object is used to track local variable references by name, while
keeping variables declared within if blocks isolated from one another.
This is only used for the parsing stage, and is discarded afterwards as
we certainly don't want expensive string lookups in the hot path.

The expression & operand parsers now take a CommandListScope object
rather than a flag indicating they are used with a command list. This
allows the operands to locate the local variables associated with the
command list.