Command List: Implement local variables
Local variables are declared like:

    local $foo = 0

Currently local variables must be explicitly initialised, though we may
relax that restriction later (particularly if we can optimise
assignments away if they are never read until the next write).

A scope object is used to track local variable references by name, while
keeping variables defined within if blocks isolated from one another.
This is only used for the parsing stage, and is discarded afterwards as
we certainly don't want expensive string lookups in the hot path.

Local variables are allocated in a stack frame associated with the
command list they are declared in, and assignments and references are
done via an offset from a stack pointer, and the stack will grow to
allocate a new stack frame whenever a command list is executed. Using a
stack in this way allows for recursive command list execution.