3DVision2SBS: Use variables instead of reserving one IniParams
Since IniParams is not namespaced it is better that we avoid reserving
any variables for shipped CustomShaders. This converts the 3dvision2sbs
shader to use a global $\ShaderFixes\3dvision2sbs.ini\mode variable for
the output mode, which it copies to an IniParams just before it runs,
backing up the original value to a local variable first and restoring it
afterwards.

Since it no longer matters which IniParams we use, switch to x0 to
minimise the IniParams buffer size.