Software Mouse Shader: Back up and restore IniParams in local variables
This adds a back up and restore mechanism to the software mouse shader
so that it won't clobber various IniParams any more, so that the
shaderhackers don't have to worry about which it might overwrite.

Since it no longer matters which ones it uses, switch to the first two
IniParams indices to avoid unnecessarily increasing the size of the
IniParams buffer.