Fix per-draw call separation / convergence
The commit:

    commit b902507222df24b2e9bf29a5bc6b215265203b9b
    Author: Ian Munsie <darkstarsword@gmail.com>
    Date:   Mon Jan 8 18:09:31 2018 +1100

        Command Lists: Introduce mechanism to set convergence from GPU

introduced a regression when setting a per-draw call separation or
convergence value that was not staged from the GPU, since it skipped
restoring the original convergence value if the staging was not
complete, which was never true when not staging the value.

This fixes the issue by setting the flag to indicate that the value was
set in the pre command list so that it will restore in the post list.

Closes #97