Warn about use of deprecated partner and depth_filter options
Partner and depth_filter options are both considered deprecated due to
some severe limitations that they poses, however they are both quite
simple options to use when appropriate and are used in a number of
fixes, so we can't simply drop support for them altogether.

However, I have recently seen two shaderhackers attempt to use partner=
in a context where it will not work, so to avoid people wasting time on
that in the future detect if partner= has been used together with a
command list and warn about it, suggesting an alternative.

We don't yet warn about partner= being used when there is no command
list since there is currently no way to specify that the original shader
should be used from a command list (handling=original?), which is
somewhat complicated by the need to hold on to a copy of the original
shader and adding that to the command list may divorce it from the
ShaderOverride and the hash that needs to be kept.

This also adds an unconditional deprecation warning for depth_filter
since texture filtering can achieve the same thing but offers more
flexibility.

These messages are issued with LOG_NOTICE priority, so they will not
show up in release mode.