3DVision2SBS: Render 3DMigoto overlay & performance monitor correctly
This adds a new command "special = draw_3dmigoto_overlay" that will draw
the 3DMigoto overlay immediately and suppresses it from being drawn in
RunFrameActions(). This is used in the SBS shader to draw the overlay
prior to the SBS pass so that the overlay will render correctly with
respect to the selected output mode.