Frame Analysis: Throw dire warning if FrameAnalysisContext is missing
I've seen a few modders enable hunting and try to use frame analysis
without restarting the game to install the FrameAnalysisContext, which
results in the FrameAnalysis directory being created, but empty except
for maybe a ShaderUsage.txt.

This adds a check when frame analysis is requested and issues a dire
warning if the context is not a FrameAnalysisContext telling the modder
to restart the game.